requirement.txt
opencv-python
ultralytics
numpy
python-socketio
aiohttp
fastapi
uvicorn

